+++ test/apis/test_api_kubernetes.py
@@ -51,12 +51,6 @@ class ApiKubernetesTest(BaseTestCase):
         self.assertEqual(200, response.status_code)
         self.assertIsNotNone(kubernetes_model)
 
-    @mock_decorator()
-    def test_put(self):
-        response = self.client.put(self.__URL)
-        self.assertEqual(200, response.status_code)
-        self.assertIsNotNone(response)
-
 
 class ApiIsKubernetesModeTest(BaseTestCase):
     __URL = f'/api/projects/{TEST_PROJECT_ID}/is_kubernetes_mode'
